Fix transfer function calculation
When the transfer function does not contains inf in denomitor or
numerator, using coeff(inf, 1) makes the transfer function undefined.
Use lcoeff(inf) instead, which uses the coeff(inf, 0) when there is no
inf in the expression, resulting the true transfer function we are
looking for.

Also delete the space column before some functions.