Add num and denom degree requirements
Sometimes the desired transfer function is expected to have degree 2
denominator, but the found topology have degree 3 denominator when the
capacitor count is increased. Avoid the degree problem by introducing
new requirements and computing new imperfections such that numerator and
denominator cannot have different degrees other than the desired
degrees.  The degree is defined as the maximum degree of an expression
of `s` here. Note that numerator and denominator are both expressions of
`s`.

This also signifies the importance of expected degree in the fitness
calculation.